Separate models from administration and put them in common
