+++ packages/common/package.json
@@ -7,6 +7,7 @@
 		"build": "tsc --build"
 	},
 	"dependencies": {
+		"@google-cloud/bigquery": "^7.8.0",
 		"cloudly-http": "^0.1.7",
 		"isly": "^0.1.16",
 		"isoly": "^2.3.11"
